bounds checks for dh_Ys and dh_Yc are always advisable
No matter what, doing a bounds-check on the peer's offered DH public
value is a good idea.  When a custom group is selected, the bounds
check is probably insufficient (there may be other problems with the
group), but it's not something either peer should ignore.

This edit clarifies that the check should happen in any case.

Thanks to Santosh Chokani for suggesting this clarification.